updated proto accuracy

+++ types/Point.go
@@ -6,14 +6,14 @@ import (
 
 // Point Stores a 2D position
 type Point struct {
-	X int32
-	Y int32
+	X float32
+	Y float32
 }
 
 //ToProto pb point to proto
 func (p *Point) ToProto() *pb.Point {
 	return &pb.Point{
-		X: p.X,
-		Y: p.Y,
+		X: float64(p.X),
+		Y: float64(p.Y),
 	}
 }
